integrating train and eval changes and removing temp files
